+++ src/Model/EmailAddress.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace CleverCloud\Sdk\Model;
+
+final readonly class EmailAddress
+{
+    public function __construct(
+        public string $email,
+        public ?bool $primary = null,
+        public ?bool $validated = null,
+    ) {
+    }
+}